benchmark param refactoring.
